remove duplicated submit_tx; move hex functions to helpers; add and cleanup comments
