+++ packages/neutron-sdk/src/interchain_queries/helpers.rs
@@ -1,5 +1,4 @@
 use crate::errors::error::{NeutronError, NeutronResult};
-use crate::interchain_queries::hex::decode_hex;
 use crate::interchain_queries::types::{
     AddressBytes, QueryPayload, QueryType, TransactionFilterItem, MAX_ADDR_LEN,
 };
@@ -9,6 +8,7 @@ use neutron_std::types::neutron::interchainqueries::{
     MsgUpdateInterchainQueryRequest,
 };
 use serde_json_wasm::to_string;
+use std::fmt::Write as _;
 
 /// Decodes a bech32 encoded string and converts to base64 encoded bytes
 /// <https://github.com/cosmos/cosmos-sdk/blob/ad9e5620fb3445c716e9de45cfcdb56e8f1745bf/types/bech32/bech32.go#L20>
@@ -47,6 +47,8 @@ pub fn uint256_to_u128(value: Uint256) -> Result<u128, StdError> {
 }
 
 /// Basic helper to define a register interchain query message:
+/// * **contract** is a contract address that registers the interchain query.
+///   Must be equal to the contract that sends the message.
 /// * **query** is a query type identifier ('tx' or 'kv' for now) with a payload:
 ///   - when the query enum is 'kv' then payload is the KV-storage keys for which we want to get
 ///     values from remote chain;
@@ -83,6 +85,8 @@ pub fn register_interchain_query(
 }
 
 /// Basic helper to define a update interchain query message:
+/// * **contract** is a contract address that updates the interchain query.
+///   Must be equal to the contract that sends the message.
 /// * **query_id** is ID of the query we want to update;
 /// * **new_keys** is encoded keys to query;
 /// * **new_update_period** is used to say how often (in neutron blocks) the query must be updated.
@@ -102,7 +106,6 @@ pub fn update_interchain_query(
             Some(filters) => {
                 to_string(&filters).map_err(|e| StdError::generic_err(e.to_string()))?
             }
-            // TODO: check if passing empty string is correct
             None => "".to_string(),
         },
     }
@@ -110,6 +113,8 @@ pub fn update_interchain_query(
 }
 
 /// Basic helper to define a remove interchain query message:
+/// * **contract** is a contract address that removes the interchain query.
+///   Must be equal to the contract that sends the message.
 /// * **query_id** is ID of the query we want to remove.
 pub fn remove_interchain_query(contract: Addr, query_id: u64) -> NeutronResult<CosmosMsg> {
     Ok(MsgRemoveInterchainQueryRequest {
@@ -131,4 +136,21 @@ pub fn kv_key_from_string<S: AsRef<str>>(s: S) -> Option<KvKey> {
         path: split[0].to_string(),
         key: decode_hex(split[1])?,
     })
+}
+
+/// Encodes bytes slice into hex string
+pub fn encode_hex(bytes: &[u8]) -> String {
+    let mut s = String::with_capacity(bytes.len() * 2);
+    for &b in bytes {
+        let _ = write!(s, "{:02x}", b);
+    }
+    s
+}
+
+/// Decodes hex string into bytes vec
+pub fn decode_hex(s: &str) -> Option<Vec<u8>> {
+    (0..s.len())
+        .step_by(2)
+        .map(|i| u8::from_str_radix(&s[i..i + 2], 16).ok())
+        .collect()
 }
