--- packages/neutron-sdk/src/interchain_queries/hex.rs
@@ -1,18 +1,0 @@
-use std::fmt::Write as _;
-
-/// Encodes bytes slice into hex string
-pub fn encode_hex(bytes: &[u8]) -> String {
-    let mut s = String::with_capacity(bytes.len() * 2);
-    for &b in bytes {
-        let _ = write!(s, "{:02x}", b);
-    }
-    s
-}
-
-/// Decodes hex string into bytes vec
-pub fn decode_hex(s: &str) -> Option<Vec<u8>> {
-    (0..s.len())
-        .step_by(2)
-        .map(|i| u8::from_str_radix(&s[i..i + 2], 16).ok())
-        .collect()
-}
