+++ packages/neutron-sdk/src/interchain_queries/mod.rs
@@ -1,5 +1,4 @@
 pub mod helpers;
-pub mod hex;
 pub mod queries;
 pub mod types;
 pub mod v045;
